requirements.txt
asgiref==3.9.1
Django==5.2.5
django-extensions==4.1
django-localflavor==5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
Markdown==3.9
PyJWT==2.10.1
python-stdnum==2.1
sqlparse==0.5.3
typing_extensions==4.15.0
